Shaked: Move loader to it's own subsystem
